+++ .gitignore
@@ -5,6 +5,7 @@
 .pydevproject
 .DS_Store
 .idea
+.vscode
 build
 install
 log
